+++ example/lib/ComplexState.dart
@@ -1,0 +1,11 @@
+class ComplexState {
+  final String title;
+  final int counterInitialValue;
+  bool isDarkTheme = false;
+
+  ComplexState({
+    this.title = "NO TITLE",
+    this.counterInitialValue = 1,
+    this.isDarkTheme = false,
+  }) {}
+}