+++ lib/src/service.dart
@@ -1,3 +1,0 @@
-abstract class Service<T> {
-  Future<List<T>> browse({String filter});
-}